Unify the sampling and the sampling-with-flow projects
Addresses second part of #82
