+++ projects/sandbox/pyproject.toml
@@ -180,6 +180,22 @@ model_state_path = "${base.basedir}/training/weights.pt"
 inference_params = "${base.inference_params}"
 basedir = "${base.basedir}"
 datadir = "${base.datadir}"
+testing_set = "${base.datadir}/flow_test_injections.h5"
+num_samples_draw = 10000
+num_plot_corner = 20
+device = "cuda"
+
+[tool.typeo.scripts.sample-and-compare-with-bilby]
+ifos = "${base.ifos}"
+kernel_length = "${base.kernel_length}"
+sample_rate = "${base.sample_rate}"
+fduration = "${base.fduration}"
+model_state_path = "${base.basedir}/training/weights.pt"
+inference_params = "${base.inference_params}"
+basedir = "${base.basedir}"
+datadir = "${base.datadir}"
+testing_set = "${base.datadir}/bilby_injections.hdf5"
+bilby_result_dir = "${base.datadir}/bilby/rundir/"
 num_samples_draw = 10000
 device = "cuda"
