--- projects/sandbox/sample_with_flow/pyproject.toml
@@ -1,34 +1,0 @@
-[tool.poetry]
-name = "sample_with_flow"
-version = "0.1.0"
-description = "Produce posterior samples of injections with our network"
-authors = ["ethan.marx <ethan.marx@ligo.org>"]
-license = "MIT"
-
-[tool.poetry.dependencies]
-python = ">=3.8,<3.11"
-
-torch = {version = "^1.12", source = "torch"}
-
-"mlpe.data" = {path = "../../../libs/data", develop  = true}
-"mlpe.trainer" = {path = "../../../libs/trainer", develop = true}
-"mlpe.logging" = {path = "../../../libs/logging", develop = true}
-"mlpe.injection" = {path = "../../../libs/injection", develop = true}
-
-"ml4gw" = {path = "../../../ml4gw", develop = true}
-typeo = {path = "../../../typeo", develop = true}
-
-ligo-skymap = "^1.0.7"
-
-[tool.poetry.scripts]
-sample-with-flow = "sample_with_flow:main"
-
-[[tool.poetry.source]]
-name = "torch"
-url = "https://download.pytorch.org/whl/cu116"
-secondary = true
-default = false
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
