+++ projects/sandbox/sampling/bilby_comparison.py
@@ -1,0 +1,112 @@
+import logging
+from pathlib import Path
+from time import time
+from typing import Callable, List
+
+from utils import (
+    add_phi_to_bilby_results,
+    draw_samples_from_model,
+    generate_corner_plots,
+    initialize_data_loader,
+    load_and_initialize_flow,
+    load_and_sort_bilby_results_from_dynesty,
+    load_preprocessor_state,
+    plot_mollview,
+)
+
+from mlpe.architectures import embeddings, flows
+from mlpe.injection.priors import sg_uniform
+from mlpe.logging import configure_logging
+from typeo import scriptify
+
+
+@scriptify(
+    flow=flows,
+    embedding=embeddings,
+)
+def main(
+    flow: Callable,
+    embedding: Callable,
+    model_state_path: Path,
+    ifos: List[str],
+    sample_rate: float,
+    kernel_length: float,
+    fduration: float,
+    inference_params: List[str],
+    datadir: Path,
+    basedir: Path,
+    testing_set: Path,
+    device: str,
+    num_samples_draw: int,
+    num_plot_corner: int,
+    bilby_result_dir: Path = None,
+    verbose: bool = False,
+):
+    logdir = basedir / "log"
+    logdir.mkdir(parents=True, exist_ok=True)
+    configure_logging(logdir / "sample_with_flow.log", verbose)
+    device = device or "cpu"
+
+    priors = sg_uniform()
+    n_ifos = len(ifos)
+    param_dim = len(inference_params)
+    strain_dim = int((kernel_length - fduration) * sample_rate)
+
+    logging.info("Initializing model and setting weights from trained state")
+    flow = load_and_initialize_flow(
+        model_state_path, embedding, n_ifos, strain_dim, param_dim, device
+    )
+    flow.eval()  # set flow to eval mode
+    logging.info(
+        "Initializing preprocessor and setting weights from trained state"
+    )
+    preprocessor_dir = basedir / "training" / "preprocessor"
+    preprocessor = load_preprocessor_state(
+        preprocessor_dir, param_dim, n_ifos, fduration, sample_rate, device
+    )
+    logging.info("Loading test data and initializing dataloader")
+    test_dataloader, params = initialize_data_loader(
+        testing_set, inference_params, device
+    )
+    logging.info(
+        "Drawing {} samples for each test data".format(num_samples_draw)
+    )
+    total_sampling_time = time()
+    results = draw_samples_from_model(
+        test_dataloader,
+        flow,
+        preprocessor,
+        inference_params,
+        num_samples_draw,
+        priors,
+        device,
+        label="test_samples_using_flow",
+    )
+    total_sampling_time = time() - total_sampling_time
+
+    logging.info("Loading bilby results")
+    bilby_results = load_and_sort_bilby_results_from_dynesty(
+        inference_params, params, bilby_result_dir
+    )
+    assert len(results) == len(bilby_results), "Length of datasets differ"
+
+    bilby_results = add_phi_to_bilby_results(bilby_results)
+
+    skymap_dir = basedir / "skymaps"
+    skymap_dir.mkdir(exist_ok=True, parents=True)
+
+    logging.info("Making joint PP plots")
+    for idx, (flow_res, bilby_res) in enumerate(zip(results, bilby_results)):
+        generate_corner_plots([flow_res, bilby_res], basedir / "corner")
+        plot_mollview(
+            flow_res.posterior["phi"].to_numpy().copy(),
+            flow_res.posterior["dec"].to_numpy().copy(),
+            truth=(bilby_res[6], bilby_res[4]),
+            outpath=skymap_dir / f"{idx}_mollview_flow.png",
+        )
+        plot_mollview(
+            bilby_res.posterior["phi"].to_numpy().copy(),
+            bilby_res.posterior["dec"].to_numpy().copy(),
+            truth=(bilby_res[6], bilby_res[4]),
+            outpath=skymap_dir / f"{idx}_mollview_bilby.png",
+        )
